+++ src/org/xjge/core/TCSetSoundMasterVolume.java
@@ -26,28 +26,28 @@ final class TCSetSoundMasterVolume extends TerminalCommand {
     }
 
     @Override
-    public void execute(List<String> args) {
+    public TerminalOutput execute(List<String> args) {
         output = null;
 
         if(args.isEmpty()) {
-            setOutput(errorNotEnoughArgs(1), Color.RED);
+            return new TerminalOutput(errorNotEnoughArgs(1), Color.RED);
         } else {
             if(args.size() > 1) {
-                setOutput(errorTooManyArgs(args.size(), 1), Color.RED);
+                return new TerminalOutput(errorTooManyArgs(args.size(), 1), Color.RED);
             } else {
                 try {
                     float value = Float.parseFloat(args.get(0));
 
                     if(value >= 0 && value <= 1) {
-                        setOutput("Sound master volume changed: (" + value + ")", Color.WHITE);
                         Audio.setSoundMasterVolume(value);
+                        return new TerminalOutput("Sound master volume changed: (" + value + ")", Color.WHITE);
                     } else {
-                        setOutput("ERROR: Value out of bounds, must be a " + 
+                        return new TerminalOutput("ERROR: Value out of bounds, must be a " + 
                                   "floating point value between 0 and 1.", 
                                   Color.RED);
                     }
                 } catch(NumberFormatException e) {
-                    setOutput(errorInvalidArg(args.get(0), "(float)"), Color.RED);
+                    return new TerminalOutput(errorInvalidArg(args.get(0), "(float)"), Color.RED);
                 }
             }
         }